Create 'Project Images' Folder and upload OrbitTest1.png
Upload of the first Orbit Test. The orbit took the shape of an oval increasing in velocity as it came closer to the larger mass, and then launching itself outward only to be pulled back in by gravity, repeating the process over and over again with about a half unit differents per revolution.