Update of Interactable Planet Trail Properties
This update of the Planet Properties changes the time active and color of the newly created trail following the object that is able to be interacted with. This allows for clearer visability when observing the changes in direction and orbits around other objects. This is the early stage development of the prediction line. 1/29/21