Implementing Teleportation Movement and Scale ratio of Planet to Player
By utilizing the XR Componets provided, Teleportation locamotion was created. These componets include XR Ray Interactor, which provides the user with a straight or curved line that is assigned a task, and Teleportation Areas and Anchors. Anchors are Platforms that only have one position when teleported to while Areas are able to be explored in the X and Z directions. The Scale of the Planets were also changed. This allows the user to see the planets interact with each other and easy access to each planet if they wish to change something (either position or velocity). Currently throwing and nudging is the only way planets can move. Implementation of Creation of planets is currently in progress and Is hope to be achieved through UI or a combination of Controller buttons that the user is able to create and scale (with its size determining how much mass it contains). Massive Walls surrounding the play area prevents planets from drifting away from the player's reach. These walls may change in the future to either be invisible or more spread out. 02/02/2021