Added Visibility Program to the Teleport Ray
Changes to the ray were made. A Program by the name LocamotionController.cs was created which only shows the ray when the button assigned to it is pressed. This allows the user to experience the program without being distracted by the ray in their vision. It grants the user to only see the ray when they click the teleport button which was set to the trigger button on the right controller. Haptic feedback was also added. When the user holds the trigger button it gives a haptic feedback of 0.1 and when teleporting to the desired location, the haptic feedback (also know as vibration) is set to 0.3. 02/03/2021