UI Implementation and Instantiation
This update brings UI into the virtual world. By using the UI objects that were built into Unity version 2019.4, buttons were built to hold the newly created planetC script. The planetC script creates a basic function that instantiates a pre-made object from the assets folder. The position and rotation were set to the left hand gameObject so when the button is pressed, the object spawns on the hand of the user as a means of easy access for grabbing and throwing. 5 buttons were created for 5 different pre-made objects. Size 1 with 0.1 mass, Size 2 with 1 mass, Size 3 with 2.5 mass, size 4 with 7 mass, and size 5 with 10 mass. These can be  infinitely instantiate. Previously in the scene have been deleted to prevent the simulation from performing without the user's permission. Displays of each object as well as they size and mass are displayed below their respective buttons. An empty button currently resides to the right of all other buttons. This was intended to be another feature but time was limited and the button has not yet been removed. The size of the play space has been increase drastically to prevent the objects from being interrupted by the walls.